requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
distlib==0.3.9
Django==5.1.5
django-debug-toolbar==5.0.1
django-filter==24.3
djangorestframework==3.15.2
drf-nested-routers==0.94.1
filelock==3.17.0
idna==3.10
mysqlclient==2.2.7
packaging==24.2
pillow==11.1.0
pipenv==2024.4.1
platformdirs==4.3.6
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
virtualenv==20.29.1
